+++ pkg/admission/prep_downscale_test.go
@@ -323,7 +323,7 @@ func testPrepDownscaleWebhook(t *testing.T, oldReplicas, newReplicas int, option
 			}, nil
 		})
 
-	admissionResponse := prepareDownscale(ctx, logger, ar, api, f)
+	admissionResponse := prepareDownscale(ctx, logger, ar, api, f, "cluster.local.")
 	require.Equal(t, params.allowed, admissionResponse.Allowed, "Unexpected result for allowed: got %v, expected %v", admissionResponse.Allowed, params.allowed)
 
 	if params.stsAnnotated {
@@ -708,6 +708,7 @@ func testPrepDownscaleWebhookWithZoneTracker(t *testing.T, oldReplicas, newRepli
 	oldRawObject, err := statefulSetTemplate(oldParams)
 	require.NoError(t, err)
 
+	clusterDomain := "cluster.local."
 	namespace := "test"
 	stsName := "my-statefulset"
 	ar := admissionv1.AdmissionReview{
@@ -794,7 +795,7 @@ func testPrepDownscaleWebhookWithZoneTracker(t *testing.T, oldReplicas, newRepli
 		}, nil
 	})
 
-	zt := newZoneTracker(api, namespace, "zone-tracker-test-cm")
+	zt := newZoneTracker(api, clusterDomain, namespace, "zone-tracker-test-cm")
 
 	admissionResponse := zt.prepareDownscale(ctx, logger, ar, api, f)
 	require.Equal(t, params.allowed, admissionResponse.Allowed, "Unexpected result for allowed: got %v, expected %v", admissionResponse.Allowed, params.allowed)
@@ -940,13 +941,14 @@ func TestCreateEndpoints(t *testing.T) {
 	}
 
 	tests := []struct {
-		name        string
-		oldInfo     *objectInfo
-		newInfo     *objectInfo
-		port        string
-		path        string
-		serviceName string
-		expected    []endpoint
+		name          string
+		oldInfo       *objectInfo
+		newInfo       *objectInfo
+		port          string
+		path          string
+		serviceName   string
+		clusterDomain string
+		expected      []endpoint
 	}{
 		{
 			name: "downscale by 2",
@@ -956,9 +958,10 @@ func TestCreateEndpoints(t *testing.T) {
 			newInfo: &objectInfo{
 				replicas: func() *int32 { i := int32(3); return &i }(),
 			},
-			port:        "8080",
-			path:        "prepare-downscale",
-			serviceName: "service-name",
+			port:          "8080",
+			path:          "prepare-downscale",
+			serviceName:   "service-name",
+			clusterDomain: "cluster.local.",
 			expected: []endpoint{
 				{
 					url:   "test-4.service-name.default.svc.cluster.local.:8080/prepare-downscale",
@@ -974,7 +977,7 @@ func TestCreateEndpoints(t *testing.T) {
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			actual := createEndpoints(ar, tt.oldInfo, tt.newInfo, tt.port, tt.path, tt.serviceName)
+			actual := createEndpoints(ar, tt.oldInfo, tt.newInfo, tt.port, tt.path, tt.serviceName, tt.clusterDomain)
 			if len(actual) != len(tt.expected) {
 				t.Errorf("createEndpoints() = %v, want %v", actual, tt.expected)
 				return